packet stuff + command confirmations

+++ src/network/packet/Packet.php
@@ -21,5 +21,5 @@ interface Packet {
 
     public function isEncoded(): bool;
 
-    public function getSentTimestamp(): ?int;
+    public function getSentTimestamp(): ?float;
 }
